Update BaseAssertionAnalyzer to create symbols at compilation start
